working...

+++ database/migrations/2022_08_02_065910_add_column_is_show_related_product_in_products_table.php
@@ -13,9 +13,11 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->boolean('is_show_related_product')->default(0);
-        });
+        if(!Schema::hasColumn('products','is_show_related_product')){
+            Schema::table('products', function (Blueprint $table) {
+                $table->boolean('is_show_related_product')->default(0);
+            });
+        }
     }
 
     /**
@@ -25,8 +27,13 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('is_show_related_product');
-        });
+
+        if(Schema::hasColumn('products','is_show_related_product')){
+            Schema::table('products', function (Blueprint $table) {
+                $table->dropColumn('is_show_related_product');
+            });
+        }
+
+
     }
 };
